+++ ocnus/src/geometry/mod.rs
@@ -15,39 +15,39 @@ pub trait OcnusGeometry<const P: usize, GS> {
     const PARAMS: [&'static str; P];
 
     /// Const array of parameter ranges.
-    const PARAM_RANGES: [(f32, f32); P];
+    const PARAM_RANGES: [(f64, f64); P];
 
     /// Compute the basis vectors in cartensian coordinates `xyz`
     /// as a function of intrinsic coordinates.
     ///
     /// These basis vectors are NOT necessarily normalized.
     fn coords_basis<CStride: Dim>(
-        ics: &VectorView3<f32>,
-        params: &VectorView<f32, Const<P>, U1, CStride>,
+        ics: &VectorView3<f64>,
+        params: &VectorView<f64, Const<P>, U1, CStride>,
         state: &GS,
-    ) -> [Vector3<f32>; 3];
+    ) -> [Vector3<f64>; 3];
 
     /// Transform cartensian coordinates `xyz` into the intrinsic coordinates.
     fn coords_ics<CStride: Dim>(
-        xyz: &VectorView3<f32>,
-        params: &VectorView<f32, Const<P>, U1, CStride>,
+        xyz: &VectorView3<f64>,
+        params: &VectorView<f64, Const<P>, U1, CStride>,
         state: &GS,
-    ) -> Vector3<f32>;
+    ) -> Vector3<f64>;
 
     /// Transform intrinsic coordinates `ics` into external cartesian coordinates.
     fn coords_xyz<CStride: Dim>(
-        ics: &VectorView3<f32>,
-        params: &VectorView<f32, Const<P>, U1, CStride>,
+        ics: &VectorView3<f64>,
+        params: &VectorView<f64, Const<P>, U1, CStride>,
         state: &GS,
-    ) -> Vector3<f32>;
+    ) -> Vector3<f64>;
 
     /// Transform a vector at/in intrinsic coordinates `ics` into external coordinates.
     fn coords_xyz_vector<CStride: Dim>(
-        ics: &VectorView3<f32>,
-        vec: &VectorView3<f32>,
-        params: &VectorView<f32, Const<P>, U1, CStride>,
+        ics: &VectorView3<f64>,
+        vec: &VectorView3<f64>,
+        params: &VectorView<f64, Const<P>, U1, CStride>,
         state: &GS,
-    ) -> Vector3<f32>;
+    ) -> Vector3<f64>;
 
     /// Retrieve a model parameter index by name.
     fn param_index(name: &str) -> usize {
@@ -61,11 +61,11 @@ pub trait OcnusGeometry<const P: usize, GS> {
     /// based on the valid range.
     ///
     /// Note that the step size for constant parameters is zero.
-    fn param_step_sizes(&self) -> [f32; P] {
+    fn param_step_sizes(&self) -> [f64; P] {
         Self::PARAM_RANGES
             .iter()
-            .map(|(min, max)| 16.0 * (max - min) * f32::EPSILON)
-            .collect::<Vec<f32>>()
+            .map(|(min, max)| 16.0 * (max - min) * f64::EPSILON)
+            .collect::<Vec<f64>>()
             .try_into()
             .unwrap()
     }
@@ -73,8 +73,8 @@ pub trait OcnusGeometry<const P: usize, GS> {
     /// Retrieve a model parameter value by name.
     fn param_value<CStride: Dim>(
         name: &str,
-        params: &VectorView<f32, Const<P>, U1, CStride>,
-    ) -> f32 {
+        params: &VectorView<f64, Const<P>, U1, CStride>,
+    ) -> f64 {
         params[Self::param_index(name)]
     }
 }
